move AppSeeAnalyticService
