+++ library/analytics/build.gradle
@@ -25,7 +25,6 @@ android {
 dependencies {
     implementation project(':library:base')
 
-    implementation "com.appsee:appsee-android:${deps.appSee.android}"
     implementation "com.crashlytics.sdk.android:crashlytics:${deps.crashlytics}"
 
     implementation "com.google.android.gms:play-services-analytics:${deps.playServices.analytics}"
